controls: Fix invalid assumptions in GravityWalker
The setCollisionRayHeight method assumed that the 0th child on the
avatar would be the av's GeomNode, and again assumed that the 2nd child of
the GeomNode would be this Walker's cRayNode. This is not always the case,
as scene graph manipulations can change the parenting order of nodes.

The fix is to simply pull the cRayNode from the Walker's own cRayNodePath
handle.