deploy-ng: Combine build_apps files and directories options into a single copy_paths
copy_paths is a list of strings and two item sequences. If a copy_paths
element is a single string, it is treated as both the source and
destination paths. If the element is a sequence, the first element is
the source and the second is the destination.

NOTE: exclude_paths now uses file globs
TODO: Support globs for copy_paths