deploy-ng: Add files to build application out of the unit tests
This will be a good way to make sure everything in the tests work under
a frozen environment.

NOTE: The application builds and runs, but no tests are collected. We
still need to figure out how we want to handle tests. Freeze them all
and modify test collection? Add py/test files to the build directory and
try to run them?