Supporting "any" state transitions in FSM
Add the possibility to transition to and from any state by using an '*'
as either Key or value in the defaultTransitions dictionary.
If '*' is given in both the key and value, any state can transition to
any other state same as if the defaultTransitions dict would be set to
None.