Revert "Display deprecation warning when writing FFT compressed anim chans"
This reverts commit 2c6a6c9f062dbb7df340133ccdce37fef2c06f2c.

This commit constitutes a commitment to keep the FFT compression feature
around in 1.10 and beyond. The plan at this time is to replace FFTW
support with FFTS (which is comparable in speed and much more liberally
licensed) and have a builtin DFT/FFT implementation to fall back on when
FFTS is not available. In this regard, FFTS does for Fourier transforms
what Eigen does for linear transforms: it provides a more efficient
implementation for Panda to use, but isn't necessary for the functionality
to work.