Have --static build a .dll as .lib no matter what. It's a static build not a mixed built.
I discovered this issue when trying to build statically in a certain way and saw the compiler trying to make a .pyd when those can't be used statically. This was not an issue before a revision to the way Python targets were added but became one after a check no longer passed for them.