gui: Add new repeatCommand argument for DirectButton
Three new arguments have been added to DirectButton: repeatCommand, repeatThreshold, and repeatDelay.

"repeatCommand" is disabled by default. If set to True, holding down a DirectButton will cause it to repeat it's command until the mouse button is released.

"repeatThreshold" is the fastest speed the repeat effect can reach. It has a default of 0.01.

"repeatDelay" is the multiplier that affects how fast the speed of the repeat effect increases. It has a default of 0.5.