roaming-ralph: Give Ralph some less-depressing skies to look at
The vertex colors of the walls in the world model were changed to be
transparent (zero alpha). This allows the window/background color to
come through. The background color has been changed from black to a sky
blue. Ideally we would use an actual sky box, but at least the sky isn't
black now.