eagldisplay: Initial implementation.
This only works on iPhones at the moment. The code is quite fragile, but
I want to have a commit that gets something on the screen that I can
build off of. A lot of the code is based off of cocoadisplay, so there
may be some unneeded stuff that I haven't removed yet.

I also want to get rid of the NSNotification method of communicating the
created view between theads. It's a bit unstable at the moment.