samples: add procedural geometry samples
This is a collection of several samples, combined with a simple sizer-based custom GUI system, based on the creation of 3D model primitives.

The samples range from very basic to quite advanced, with a couple of demo's intended for pure entertainment thrown in.
Apart from a simple cube, five standard 3D model primitive types can be generated with the code provided in the src/prim package:
box, sphere, cylinder, cone and torus. Each of these can be created with a plethora of options, such that the resulting models can vary dramatically in shape and UV-mapping.