multitouch: Refactor PointerData and its usage to better support multitouch.
The goal is to have InputDevice be authoratative over the *current* state of each pointer instead of different copies of PointerData flying around. Instead, each PointerData will be modified directly, and will only be copied when being packed into a PointerEvent (which will be changed to hold only PointerData and a timestamp).