makepackage: Change macOS install location
Starting with macOS 10.15, we can no longer install into
/Developer/Panda3D due to it not being one of the sanctioned writable
directories. This commit changes the install location to
/Library/Developer/Panda3D, which is the next best place
to put it (discussion on this can be found in #760).

To keep compatibility with those who may rely on the old location, some
postinstall scripts have been added that create symlinks to the new
location if running on 10.14 or older.

Fixes #760.