shaderpipeline: SPIR-V improvements, fixes for shadow samplers
Shadow map samplers weren't being marked as such by glslang if they were declared as a function parameter.  This is arguably a glslang bug, but the spec is vague enough about whether it's valid or not to specify this that it's worth fixing this on the Panda end.

As an added bonus, we now have a basic level of usage tracking for variables, so we can implement unused variable elimination if we wanted to.