saving copies of openal implementation
Since I'm not the most experienced in C++, I'm wondering if I should try to write my implementation from scratch while only using the openAL implementation as a reference. To be safe, I'm saving the files I made out to be copies so I can try modifying the tried-and-true implementation later if need be.