pgraphnodes: Disable lighting when rendering shadow cam
This prevents accessing a texture (the shadow map) that we're currently rendering into.  Hopefully a reasonable driver was already optimizing out the access properly due to the color write being off, but it could cause unnecessary work in the driver

This will have the effect of the shader generator generating a different shader for the shadow pass than for the main pass, ideally we optimize the shader generator a bit more so that there are fewer shaders generated for the shadow pass.