CutScene Class for Direct
Issue description

People often times want to be able to wait for a event in-between intervals.

Solution description

I have created a CutScene class which holds a list of Intervals and WaitFor classes. The CutScene class goes through each interval and if the next one is a WaitFor object it won't progress until either the event is fired or (if supplied) the WaitFor 'times out'. I'd say that this class is pretty simple but should speed up the workflow of direct users.

Checklist

I have done my best to ensure that…

…I have familiarized myself with the CONTRIBUTING.md file
…this change follows the coding style and design patterns of the codebase
…I own the intellectual property rights to this code
…the intent of this change is clearly explained
…existing uses of the Panda3D API are not broken
…the changed code is adequately covered by the test suite, where possible.